+++ public/packages/types/types/config/app.d.ts
@@ -1,8 +1,17 @@
 export interface App {
-  // URL under which the app will be available locally
-  localURL?: string,
-  // Ports to be exposed
-  ports?: string[]
   // Name of task to be run
   task: string
+  // Expose ports for the app
+  expose: Expose[]
+}
+
+interface Expose {
+  // URL under which your app will be available locally on your machine
+  localHost?: string
+  // Port where the app runs
+  port?: number
+  // Access sets the bit-ship network access
+  // - public: app port is exposed to host machine for example localhost:8080
+  // - proxy-public: app expose port only for the proxy server that runs under the host machine localhost:80
+  access? : 'public' | 'proxy-public'
 }